Modify actionpolicy.rb to enable a default permissions file. This works based on two (optional) config file settings:
actionpolicy.enable_default
actionpolicy.default_name

If the former is set to 1, actionpolicy.rb will first look for a <pluginname>.policy as per
the original actionpolicy.rb.  If this lookup fails it will next look for the following:

If there is no actionpolicy.default_name setting in the config file, we look for a file
named default.policy

If there is a default_name setting, we look for <default_name>.policy

In either case we then substitute this for <pluginname>.policy and process.