(#8955) Refactor puppetral agent
This refactor adds the ability to find and search with the puppetral
agent. The "do" action is renamed to "create" to be more consistent
with the "puppet resource" commandline tool.

We added specs for the puppetral agent.

The find and search output format can now be piped to the create
action.

Reviewed by: Matt Robinson <matt@puppetlabs.com>, Joshua Harlan Lifton
<lifton@puppetlabs.com>, and R.I. Pienaar <rip@devco.net>.