Fix bad input stub => failing test for puppetd agent.
In Ruby you can use an array deference on a Fixnum and get the bits at that
position; when you use a non-integer key you get zero, which is valid.

The test stubbed 0, a Fixnum, as a return value to stimulate an error return,
but it actually worked.  Replacing that with a more invalid value lets us get
the failure expected, and so exercise the right code path.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>