(#9923) Support puppet agent as a daemon from puppetd
The `puppetd` MCollective agent didn't play nicely with a daemonized Puppet
agent; it would notice the lockfile and refuse to run.  Now, instead, it will
send SIGUSR1 to the daemon.

That will trigger an immediate run of Puppet, unless we are still in the
middle of the first (and only) splay interval.  During that window the agent
will just pick a new splay and wait that out; this is a limitation of the
agent and there is nothing else we can do.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>