Retain provider when searching packages
Recent changes to the agent caused all non-property parameters to be
trimmed from resources when finding & searching; this is desirable in
all cases except for when dealing with packages, where it is necessary
to know the provider. This commit adds a method for parameter pruning
(which was previously being duplicated between the find and search
action) and also special-cases package providers to retain them.

Reviewed-by: Joshua Lifton <lifton@puppetlabs.com>