Change puppetral#create to return more resource info
Instead of returning the result of the Puppet::Resource.indirection.save
under the :resource key, we now instead do a find. This has the effect
of returning more information about the created resource, since the
return value of indirection.save only contains info about the parameters
that changed, and not about all properties of the resource.

Reviewed-by: Joshua Lifton <lifton@puppetlabs.com>