Fix order dependent test failures
running the package spec before the puppetral spec got the following
problem:

    Failure/Error: @agent = MCollective::Test::LocalAgentTest.new("puppetral", :agent_file => agent_file).plugin
    RuntimeError:
      Plugin connector_plugin already loaded
    # /home/vagrant/work/marionette-collective/lib/mcollective/pluginmanager.rb:37:in `<<'
    # /home/vagrant/work/mcollective-test/lib/mcollective/test/util.rb:60:in `create_connector_mock'
    # /home/vagrant/work/mcollective-test/lib/mcollective/test/local_agent_test.rb:15:in `initialize'
    # ./agent/puppetral/spec/puppetral_spec.rb:7:in `new'
    # ./agent/puppetral/spec/puppetral_spec.rb:7

For some reason the PluginManager isn't clearing properly in the package specs
with the spec_helper.  After manually clearing the PluginManager there was
still the issue of the package test reopening the Puppet::Type, which then also
caused problems for puppetral.

Now all the specs pass with the rake task.