Maint: Fix "name" field for long help output
MCollective's long help output uses the "name" field as the main header for an
agent's help page, which makes the short name (as used when invoking)
impractical for use in that field. This commit changes the name fields of the
package, service, puppetd, and puppetral agents/ddl files to short
descriptions suitable for long help output.