(#10221) Update puppetd plugin to take advantage of new agent statuses
Ticket #9923 allowed the puppetd plugin to trigger agent runs even when
the agent was already daemonized, but it introduced some new state
information about the puppet agent, :idling and :stopped, that wasn't
consolidated very well with the current understanding of puppet state by
other areas on the plugin.

This commit consolidates that state information to reply[:status], and
adds the reply[:idling] and reply[:stopped], like the reply[:running]
and reply[:enabled], for backward compatibility and easy command line
manipulation of the status information.

The agent, application and ddl were all updated to take advantage of
this new state information.  Also bumped the version from 1.4 to 1.5.