(#10221) Update puppetd plugin to take advantage of new agent statuses
Ticket #9923 allowed the puppetd plugin to trigger agent runs even when
the agent was already daemonized, but it introduced some new state
information about the puppet agent, :idling and :stopped, that wasn't
consolidated very well with the current understanding of puppet state by
other areas on the plugin.

This commit consolidates that state information to reply[:status], but
leaves the reply[:running] and reply[:enabled] for backward
compatibility.  This involves some minor rewording to messages returned
by the application and some code refactors.

Paired-with: Patrick Carlisle <patrick@puppetlabs.com>