(#10221) Readd status to puppetd status output
The output used to be a string with a summary of all the info you would
want about the status of the puppet agent, but with the rework of the
status to including more states, the state information was moved out of
the output.  This puts it back in for a friendlier status string.